Moved API logic out of TraitToggle
